+++ src/test/java/com/rafa/autorizador/cartao/saldo/BuscarSaldoServiceTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -14,9 +15,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.rafa.autorizador.cartao.Cartao;
-import com.rafa.autorizador.cartao.CartaoRepository;
-import com.rafa.autorizador.config.exception.CartaoInexistenteException;
+import com.rafa.autorizador.cartao.exception.modelo.CartaoInexistenteException;
+import com.rafa.autorizador.cartao.modelo.Cartao;
+import com.rafa.autorizador.cartao.modelo.CartaoRepository;
 
 @ExtendWith(MockitoExtension.class)
 public class BuscarSaldoServiceTest {
@@ -28,15 +29,15 @@ public class BuscarSaldoServiceTest {
     @BeforeEach
     public void setUp() {
         numeroCartao = "1616161616161616";
-        cartao = Cartao.builder().numeroCartao(numeroCartao).saldo(500.0).build();
+        cartao = Cartao.builder().numeroCartao(numeroCartao).saldo(BigDecimal.valueOf(500.00)).build();
     }
 
     @Test
     public void testBuscaSaldo_Success() throws CartaoInexistenteException {
         when(cartaoRepository.findByNumeroCartao(anyString())).thenReturn(Optional.of(cartao));
 
-        Double saldo = buscarSaldoService.busca(numeroCartao);
-        assertEquals(500.0, saldo);
+        BigDecimal saldo = buscarSaldoService.busca(numeroCartao);
+        assertEquals(BigDecimal.valueOf(500.00), saldo);
     }
 
     @Test
